同步更新代码

+++ service/buffer.go
@@ -126,7 +126,7 @@ func (this *buffer) Close() error {
 	atomic.StoreInt64(&this.done, 1)
 
 	this.pcond.L.Lock()
-	for i := 0; i < this.size; i++ {
+	for i := int64(0); i < this.size; i++ {
 		this.buf[i] = nil
 	}
 	this.pcond.L.Unlock()
